# [BugFix] Optimize mv relatd locks