… fe/fe-core/src/main/java/com/starrocks/scheduler/mv/MVVersionManager.java
… fe/fe-core/src/main/java/com/starrocks/server/LocalMetastore.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MvRewritePreprocessor.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvUtils.java
… fe/fe-core/src/test/java/com/starrocks/common/lock/TestLockInterface.java