## Why I'm doing:
Bug 1: `updateMetaForOlapTable` is slow:
```
      {
         "id":9658056,
         "name":"starrocks-taskrun-pool-26084",
         "type":"INTENTION_EXCLUSIVE",
         "heldFor":6980,
         "waitTime":0,
         "stack":[
            "java.base@11.0.28/java.util.HashMap.hash(HashMap.java:340)",
            "java.base@11.0.28/java.util.HashMap.put(HashMap.java:608)",
            "java.base@11.0.28/java.util.HashSet.add(HashSet.java:220)",
            "java.base@11.0.28/java.util.stream.Collectors$$Lambda$35/0x00000008000fd040.accept(Unknown Source)",
            "java.base@11.0.28/java.util.stream.ReduceOps$3ReducingSink.accept(ReduceOps.java:169)",
            "java.base@11.0.28/java.util.stream.ReferencePipeline$2$1.accept(ReferencePipeline.java:177)",
            "java.base@11.0.28/java.util.TreeMap$KeySpliterator.forEachRemaining(TreeMap.java:2739)",
            "java.base@11.0.28/java.util.stream.AbstractPipeline.copyInto(AbstractPipeline.java:484)",
            "java.base@11.0.28/java.util.stream.AbstractPipeline.wrapAndCopyInto(AbstractPipeline.java:474)",
            "java.base@11.0.28/java.util.stream.ReduceOps$ReduceOp.evaluateSequential(ReduceOps.java:913)",
            "java.base@11.0.28/java.util.stream.AbstractPipeline.evaluate(AbstractPipeline.java:234)",
            "java.base@11.0.28/java.util.stream.ReferencePipeline.collect(ReferencePipeline.java:578)",
            "app//com.starrocks.catalog.OlapTable.getVisiblePartitionNames(OlapTable.java:1680)",
            "app//com.starrocks.scheduler.mv.MVVersionManager.lambda$updateMetaForOlapTable$2(MVVersionManager.java:123)",
            "app//com.starrocks.scheduler.mv.MVVersionManager$$Lambda$3757/0x00000008014bd040.test(Unknown Source)",
            "java.base@11.0.28/java.util.Collection.removeIf(Collection.java:544)",
            "app//com.starrocks.scheduler.mv.MVVersionManager.updateMetaForOlapTable(MVVersionManager.java:122)",
            "app//com.starrocks.scheduler.mv.MVVersionManager.updateMVVersionInfo(MVVersionManager.java:73)",
            "app//com.starrocks.scheduler.PartitionBasedMvRefreshProcessor.updateMeta(PartitionBasedMvRefreshProcessor.java:831)",
            "app//com.starrocks.scheduler.PartitionBasedMvRefreshProcessor.doRefreshMaterializedView(PartitionBasedMvRefreshProcessor.java:472)",
            "app//com.starrocks.scheduler.PartitionBasedMvRefreshProcessor.doRefreshMaterializedViewWithRetry(PartitionBasedMvRefreshProcessor.java:376)",
            "app//com.starrocks.scheduler.PartitionBasedMvRefreshProcessor.doMvRefresh(PartitionBasedMvRefreshProcessor.java:335)",
            "app//com.starrocks.scheduler.PartitionBasedMvRefreshProcessor.processTaskRun(PartitionBasedMvRefreshProcessor.java:203)",
            "app//com.starrocks.scheduler.TaskRun.executeTaskRun(TaskRun.java:313)",
            "app//com.starrocks.scheduler.TaskRunExecutor.lambda$executeTaskRun$0(TaskRunExecutor.java:60)",
            "app//com.starrocks.scheduler.TaskRunExecutor$$Lambda$2841/0x000000080145d840.get(Unknown Source)",
            "java.base@11.0.28/java.util.concurrent.CompletableFuture$AsyncSupply.run(CompletableFuture.java:1700)",
            "java.base@11.0.28/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1128)",
            "java.base@11.0.28/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:628)",
            "java.base@11.0.28/java.lang.Thread.run(Thread.java:829)"
         ]
      }
```
Problem 2: Query's `Lock` time may be too long if it contains related mvs:
```
  Planner:
     - -- Parser[1] 0
     - -- Total[1] 6s883ms
     -     -- Analyzer[1] 6s732ms
     -         -- Lock[1] 6s725ms
     -         -- AnalyzeDatabase[3] 0
     -         -- AnalyzeTemporaryTable[3] 0
     -         -- AnalyzeTable[3] 0
     -     -- Transformer[1] 2ms
     -     -- Optimizer[1] 144ms
     -         -- MVPreprocess[1] 1ms
     -             -- MVChooseCandidates[1] 0
     -         -- MVTextRewrite[1] 0
     -         -- RuleBaseOptimize[1] 92ms
     -         -- CostBaseOptimize[1] 40ms
     -         -- PhysicalRewrite[1] 10ms
     -         -- PlanValidate[1] 0
     -             -- InputDependenciesChecker[1] 0
     -             -- TypeChecker[1] 0
     -             -- CTEUniqueChecker[1] 0
     -             -- ColumnReuseChecker[1] 0
     -     -- ExecPlanBuild[1] 0
     - -- Pending[1] 0
     - -- Prepare[1] 0
     - -- Deploy[1] 7ms
     -     -- DeployLockInternalTime[1] 7ms
     -         -- DeploySerializeConcurrencyTime[3] 0
     -         -- DeployStageByStageTime[9] 0
     -         -- DeployWaitTime[9] 5ms
     -             -- DeployAsyncSendTime[4] 0
     - DeployDataSize: 23604
```
## What I'm doing:
- Ensure mv refresh should own `database` level write lock to avoid affecting queries;
- Optimize `updateMetaForOlapTable` for better performance.

Fixes #issue

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 4.0
  - [x] 3.5
  - [x] 3.4
  - [x] 3.3